Implement authorization for restricted delete APIs (Added ownership-based and admin-level authorization checks to ensure that only the resource creator or an admin can delete the resource.)

+++ src/main/java/com/example/Quora/Services/AnswerService.java
@@ -8,7 +8,9 @@ import com.example.Quora.Entities.Answer;
 import com.example.Quora.Entities.Question;
 import com.example.Quora.Exceptions.AnswerNotFoundException;
 import com.example.Quora.Exceptions.InvalidInputException;
+import com.example.Quora.Exceptions.UnauthorizedException;
 import com.example.Quora.Repositories.AnswerRepository;
+import com.example.Quora.Repositories.QuestionRepository;
 
 import io.micrometer.common.util.StringUtils;
 
@@ -19,7 +21,13 @@ public class AnswerService {
 	private AnswerRepository answerRepository;
 
 	@Autowired
+	private QuestionRepository questionRepository;
+
+	@Autowired
 	private QuestionService questionService;
+
+	@Autowired
+	private UserService userService;
 
 	public Answer createNewAnswer(final AnswerDto ans) throws InvalidInputException {
 		if (ans == null || StringUtils.isBlank(ans.getAnswer())) {
@@ -29,7 +37,7 @@ public class AnswerService {
 		final Question question = questionService.getQuestionByQuestionId(ans.getQId());
 
 		final Answer answer = Answer.builder().answer(ans.getAnswer()).createdBy(ans.getCreatedBy())
-				.modifiedBy(ans.getModofiedBy()).question(question).build();
+				.modifiedBy(ans.getModifiedBy()).question(question).build();
 
 		final Answer newAnswer = answerRepository.save(answer);
 
@@ -49,8 +57,16 @@ public class AnswerService {
 
 	public Answer deleteAnswer(final int ansId) throws InvalidInputException, AnswerNotFoundException {
 
-		final Answer existingAns = getAnswerByAnswerId(ansId);
+		final String loggedInUsername = userService.getLoggedInUserName();
+		final boolean isAdmin = userService.isloggedInUserAdmin();
 
+		final Answer existingAns = getAnswerByAnswerId(ansId);
+		final String questionCreatedBy = questionRepository.findCreatedBy(existingAns.getQuestion().getId());
+
+		if (!(loggedInUsername.equals(existingAns.getCreatedBy()) || loggedInUsername.equals(questionCreatedBy)
+				|| isAdmin)) {
+			throw new UnauthorizedException("You are not authorized to perform delete operation.");
+		}
 		answerRepository.deleteById(ansId);
 
 		return existingAns;
