+++ src/main/java/com/example/Quora/Services/CommentService.java
@@ -8,6 +8,8 @@ import com.example.Quora.Entities.Answer;
 import com.example.Quora.Entities.Comment;
 import com.example.Quora.Exceptions.CommentNotFoundException;
 import com.example.Quora.Exceptions.InvalidInputException;
+import com.example.Quora.Exceptions.UnauthorizedException;
+import com.example.Quora.Repositories.AnswerRepository;
 import com.example.Quora.Repositories.CommentRepository;
 
 import io.micrometer.common.util.StringUtils;
@@ -20,6 +22,12 @@ public class CommentService {
 
 	@Autowired
 	private AnswerService answerService;
+
+	@Autowired
+	private UserService userService;
+
+	@Autowired
+	private AnswerRepository answerRepository;
 
 	public Comment createComment(final CommentDto commentDto) throws InvalidInputException {
 		if (commentDto.getComment() == null || StringUtils.isBlank(commentDto.getComment())) {
@@ -48,9 +56,17 @@ public class CommentService {
 	}
 
 	public Comment deleteComment(final int commentId) throws InvalidInputException, CommentNotFoundException {
+		final String loggedInUsername = userService.getLoggedInUserName();
+		final boolean isAdmin = userService.isloggedInUserAdmin();
 
 		final Comment existingComment = getCommentByCommentId(commentId);
 
+		final String answerCreatedBy = answerRepository.findCreatedBy(existingComment.getAnswer().getId());
+
+		if (!(loggedInUsername.equals(existingComment.getCreatedBy()) || loggedInUsername.equals(answerCreatedBy)
+				|| isAdmin)) {
+			throw new UnauthorizedException("You are not authorised to perform delete operation.");
+		}
 		commentRepository.deleteById(commentId);
 
 		return existingComment;
